requirements.txt
Flask
Flask-PyMongo
flask-mongo-sessions
pyes
Werkzeug
ordereddict
datetime
python-dateutil
lxml
translitcodec
pytz
cssselect
requests
Flask-Basicauth
Flask-WTF
Flask-Bootstrap
Flask-Cache
elasticsearch
numpy
matplotlib
imposm.parser
python-memcached
pillow
git+git://github.com/werner2101/python-osm.git
